Refactoring - use ? instead of match
